moved xor to falcon_tensor.py
