+++ test/torch/mpc/falcon/test_falcon_tensor
@@ -1,0 +1,23 @@
+import torch
+
+
+def test_public_xor(workers):
+    bob, alice, james = (workers["bob"], workers["alice"], workers["james"])
+    x = torch.tensor([0]).share(bob, alice, james, protocol="falcon", field=2)
+    y = torch.tensor([1])
+    a = torch.tensor([1]).share(bob, alice, james, protocol="falcon", field=2)
+    b = torch.tensor([0])
+    assert (x.xor(y).reconstruct() == torch.tensor(1)).all()
+    assert (x.xor(b).reconstruct() == torch.tensor(0)).all()
+    assert (a.xor(y).reconstruct() == torch.tensor(0)).all()
+    assert (a.xor(b).reconstruct() == torch.tensor(1)).all()
+
+
+def test_private_xor(workers):
+    bob, alice, james = (workers["bob"], workers["alice"], workers["james"])
+    x = torch.tensor([0]).share(bob, alice, james, protocol="falcon", field=2)
+    y = torch.tensor([1]).share(bob, alice, james, protocol="falcon", field=2)
+    assert (x.xor(y).reconstruct() == torch.tensor(1)).all()
+    assert (x.xor(x).reconstruct() == torch.tensor(0)).all()
+    assert (y.xor(y).reconstruct() == torch.tensor(0)).all()
+    assert (y.xor(x).reconstruct() == torch.tensor(1)).all()
