+++ test/torch/tensors/test_replicated_shared.py
@@ -7,7 +7,7 @@ def test_sharing(workers):
     bob, alice, james = (workers["bob"], workers["alice"], workers["james"])
     plain_text = torch.tensor([3, 7, 11])
     secret = plain_text.share(bob, alice, james, protocol="falcon")
-    assert type(secret) == syft.ReplicatedSharingTensor
+    assert isinstance(secret, syft.ReplicatedSharingTensor)
     assert type(secret.child) == dict
 
 
@@ -101,28 +101,6 @@ def test_view(workers):
     x = torch.rand([2, 1]).share(bob, alice, james, protocol="falcon")
     x = x.view([1, 2])
     assert x.shape == torch.Size([1, 2])
-
-
-def test_public_xor(workers):
-    bob, alice, james = (workers["bob"], workers["alice"], workers["james"])
-    x = torch.tensor([0]).share(bob, alice, james, protocol="falcon", field=2)
-    y = torch.tensor([1])
-    a = torch.tensor([1]).share(bob, alice, james, protocol="falcon", field=2)
-    b = torch.tensor([0])
-    assert (x.xor(y).reconstruct() == torch.tensor(1)).all()
-    assert (x.xor(b).reconstruct() == torch.tensor(0)).all()
-    assert (a.xor(y).reconstruct() == torch.tensor(0)).all()
-    assert (a.xor(b).reconstruct() == torch.tensor(1)).all()
-
-
-def test_private_xor(workers):
-    bob, alice, james = (workers["bob"], workers["alice"], workers["james"])
-    x = torch.tensor([0]).share(bob, alice, james, protocol="falcon", field=2)
-    y = torch.tensor([1]).share(bob, alice, james, protocol="falcon", field=2)
-    assert (x.xor(y).reconstruct() == torch.tensor(1)).all()
-    assert (x.xor(x).reconstruct() == torch.tensor(0)).all()
-    assert (y.xor(y).reconstruct() == torch.tensor(0)).all()
-    assert (y.xor(x).reconstruct() == torch.tensor(1)).all()
 
 
 # corner cases
